requirements.txt
yara-python
bcrypt
pywin32
wmi
pywebview
psutil
Flask
